requirements.txt
 asttokens==2.4.1
 attrs==23.1.0
 beautifulsoup4==4.12.2
 certifi==2023.7.22
 cffi==1.16.0
 colorama==0.4.6
 comm==0.2.0
 debugpy==1.8.0
 decorator==5.1.1
 executing==2.0.1
 h11==0.14.0
 html5lib==1.1
 idna==3.4
 ipykernel==6.26.0
 ipython==8.17.2
 jedi==0.19.1
 jupyter_client==8.6.0
 jupyter_core==5.5.0
 lxml==4.9.3
 matplotlib-inline==0.1.6
 nest-asyncio==1.5.8
 numpy==1.26.2
 outcome==1.3.0.post0
 packaging==23.2
 pandas==2.1.3
 parso==0.8.3
 platformdirs==4.0.0
 prompt-toolkit==3.0.41
 psutil==5.9.6
 pure-eval==0.2.2
 pycparser==2.21
 Pygments==2.16.1
 PySocks==1.7.1
 python-dateutil==2.8.2
 pytz==2023.3.post1
 pywin32==306
 pyzmq==25.1.1
 selenium==4.15.2
 webdriver-manager==3.8.6
 beautifulsoup4==4.12.2
 lxml==4.9.3
 requests>=2.31.0
 SQLAlchemy>=2.0.20
 python-dotenv>=1.0.0
 rapidfuzz>=2.15.1
 streamlit>=1.0.0
 altair>=4.0.0
 matplotlib>=3.0.0
 scipy>=1.10.0
 scikit-learn>=1.0.0